Now checks the first,last,and middle initial as well as the ID number
I know we weren't going to do this origionally but this way it's more secure and if someone gets your VRN, DLN, or SSN and doesn't have the name then they can't log in. If we take the fields we might as well test them you know? Als  I added skeleton methods for chnging the "has voted" variable later but didn't really do anything with them.