Added calls to the  setHasVoted and beginElection methods in the right places
The textInterface now calls the two methods at the bottom that will eventually set the "has Voted" colun of the voters in the registered. file. The beginElection sets all voter's to false becuase they haven't voted in this specific election yet. The sethasVoted sets the specific voter passed to true so that they can't vote again in this election.